requirements.txt
streamlit
pytesseract
pillow
numpy
sympy
opencv-python
